requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
twilio
openai
google-api-python-client
